% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ubigeos}
\alias{ubigeos}
\title{Peruvian Ubigeos.}
\format{A data frame with 1876 rows and 18 variables}
\source{
\url{https://github.com/CONCYTEC/ubigeo-peru/blob/master/equivalencia-ubigeos-oti-concytec.csv}
}
\usage{
ubigeos
}
\description{
A dataset containing three peruvian ubigeos clasification from INEI, RENIEC and SUNAT.
}
\keyword{datasets}
