% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubigeos_tidy.R
\docType{data}
\name{ubigeos_tidy}
\alias{ubigeos_tidy}
\title{Peruvian Ubigeos (tidy version).}
\format{A data frame with 6206 rows and 4 variables}
\source{
\url{https://github.com/CONCYTEC/ubigeo-peru/blob/master/equivalencia-ubigeos-oti-concytec.csv}
}
\usage{
ubigeos_tidy
}
\description{
This is the tidy version of the ubigeos dataset. A dataset containing three
peruvian ubigeos clasification from INEI, RENIEC and SUNAT.
}
\keyword{datasets}
